pyproject.toml
[project]
name = "voice-assistant-demo"
version = "0.1.0"
description = "Customer support hotline with Twilio + Google ADK (Gemini Live API)"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.28.0",
    "twilio>=9.0.0",
    "audioop-lts>=0.2.1",   # mulaw <-> pcm conversion (Python 3.13+ compatible)
]

[tool.pytest.ini_options]
filterwarnings = [
    # silence as long as https://github.com/googleapis/python-genai/pull/1939 is not merged
    "ignore:'_UnionGenericAlias' is deprecated:DeprecationWarning",
]

[tool.coverage.report]
skip_covered = true

[tool.uv]
package = true

[tool.uv.sources]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
]
